* Extracted the VisLog's content storage and writing methods to a separate class PageSession which shall prepare the session and user specific updating of content
