+++ scistag/tests/vislog/test_pandas_logger.py
@@ -5,7 +5,6 @@ Tests the pandas specific logging functions
 import shutil
 from unittest import mock
 
-import numpy as np
 import pytest
 
 from . import vl
