+++ scistag/tests/vislog/test_time_logger.py
@@ -13,11 +13,11 @@ def test_time_logging_basics():
     test_log = VisualLog(log_to_disk=False)
     test_log.default_builder.time.log()
     test_log.default_builder.time().flush()
-    body = test_log.get_body("html")
+    body = test_log.default_page.get_body("html")
     assert (
         b"-" in body
         and b":" in body
         and b"." in body
         and body.index(b".") > body.index(b":")
     )
-    vl.embed(test_log)
+    vl.embed(test_log.default_page)
