+++ scistag/vislog/extensions/pandas_logger.py
@@ -87,4 +87,4 @@ class PandasLogger(BuilderExtension):
             else:
                 self.builder.add_md(string_table)
             self.builder.add_txt(string_table)
-        self.builder.target_log.handle_modified()
+        self.page.handle_modified()
