+++ scistag/vislog/sessions/__init__.py
@@ -1,0 +1,3 @@
+"""
+Defines the single session data types such as page and user sessions
+"""
