+++ scistag/vislog/visual_log.py
@@ -12,11 +12,9 @@ from typing import TYPE_CHECKING, Callable, Union, Type, Literal
 from scistag.common import StagLock, Cache, StagApp
 from scistag.filestag import FileStag, FilePath
 from scistag.imagestag import Size2D, Size2DTypes
-from scistag.vislog.common.visual_log_element import VisualLogElement
 from scistag.webstag.web_helper import WebHelper
 from scistag.logstag.console_stag import Console
-
-ROOT_ELEMENT_NAME = "body"
+from scistag.vislog.sessions.page_session import PageSession, HTML, MD, TXT, CONSOLE
 
 if TYPE_CHECKING:
     from scistag.webstag.server import WebStagServer
@@ -77,16 +75,6 @@ LOG_EVENT_CACHE_NAME = "__logEvents"
 "Name of the cache entry in which the log events are stored"
 
 SLEEPING_TEXT_MESSAGE = "\nZzzzzzz - Press Ctrl-C to terminate\n"
-
-# Definition of output types
-HTML = "html"
-"Html output"
-CONSOLE = "console"
-"Console output"
-TXT = "txt"
-"Txt output"
-MD = "md"
-"Markdown output"
 
 CUTE_APP = "cute"
 "Defines that the app shall be started in a Qt browser window"
@@ -157,7 +145,7 @@ class VisualLog:
         """
         :param target_dir: The output directory
         :param title: The log's name
-        :param index: Defines the name of the log's index file. index by
+        :param index_name: Defines the name of the log's index file. index by
             default.
 
             Extensions such as .html, .md and .txt will automatically
@@ -224,10 +212,10 @@ class VisualLog:
         :param auto_reload: Defines if this log will be executed in auto_reload
             mode in its cache should be update and restored each turn.
         """
-        self._page_lock = StagLock()
-        "Lock for multithread secure access to the latest page update"
-        self._backup_lock = StagLock()
-        "Lock for multithread secure access to the latest page update backup"
+        self._general_lock = StagLock()
+        """
+        Access lock to handle the page's events
+        """
         self.start_time = time.time()
         "The time stamp of when the log was creation"
         try:
@@ -236,7 +224,11 @@ class VisualLog:
         except FileNotFoundError:
             pass
         if formats_out is None:
-            formats_out = {"html"}
+            formats_out = {HTML}
+        self.index_name = index_name
+        """
+        The index file's name
+        """
         self._app = app
         "Defines the initial application to wind up with this log"
         self._resolution = resolution
@@ -294,20 +286,8 @@ class VisualLog:
         self.log_formats = formats_out
         "Defines if text shall be logged"
         self.log_formats.add(CONSOLE)
-        self._logs = VisualLogElement(
-            name=ROOT_ELEMENT_NAME, output_formats=sorted(self.log_formats)
-        )
-        self._log_backup: VisualLogElement | None = None
-        """
-        Defines a copy of the current data
-        """
-        self.create_log_backup()
         """
         Contains the log data for each output type
-        """
-        self.cur_element = self._logs
-        """
-        Defines the current target element to which new data shall be added
         """
         self.continuous_write = continuous_write
         "If defined the output logs will be updated after every log"
@@ -325,20 +305,6 @@ class VisualLog:
         "The default image type to use for storage"
         self.image_quality = image_quality
         "The image compression quality"
-        self._html_export = HTML in self.log_formats
-        "Defines if HTML gets exported"
-        self.md_export = MD in self.log_formats
-        "Defines if markdown gets exported"
-        self.txt_export = TXT in self.log_formats
-        "Defines if txt gets exported"
-        self.index_name = index_name
-        "Defines the log's index file name"
-        self._txt_filename = self.target_dir + f"/{self.index_name}.txt"
-        "The name of the txt file to which we shall save"
-        self._html_filename = self.target_dir + f"/{self.index_name}.html"
-        "The name of the html file to which we shall save"
-        self._md_filename = self.target_dir + f"/{self.index_name}.md"
-        "The name of the markdown file to which we shall save"
         self._consoles: list[Console] = []
         "Attached consoles to which the data shall be logged"
         from scistag.vislog.renderers.log_renderer_html import HtmlLogRenderer
@@ -347,15 +313,6 @@ class VisualLog:
             HTML: HtmlLogRenderer(title=self._title)
         }
         "The renderers for the single supported formats"
-        self._page_backups: dict[str, bytes] = {}
-        """
-        A backup of the latest rendered page of each dynamic data type
-        (excluding PDFs and PNGs which are just created on demand)
-        """
-        self._body_backups: dict[str, bytes] = {}
-        """
-        A backup of the latest body renderings for each content type
-        """
         self.static_files: dict[str, bytes] = {}
         "Statically hosted files for a pure web based provision of the log"
         self._shall_terminate = False
@@ -377,10 +334,6 @@ class VisualLog:
             self._events = self.cache.get(LOG_EVENT_CACHE_NAME, default=[])
         self._invalid = False
         "Defines if this log was invalidated via :meth:`invalidate`"
-        self.name_counter = Counter()
-        "Counter for file names to prevent writing to the same file twice"
-        self.title_counter = Counter()
-        "Counter for titles to numerate the if appearing twice"
         self._total_update_counter = 0
         "The total number of updates to this log"
         self._update_counter = 0
@@ -391,7 +344,19 @@ class VisualLog:
         # The last computed updated rate in updates per second
         from .visual_log_builder import VisualLogBuilder
 
-        self.default_builder: VisualLogBuilder = VisualLogBuilder(self)
+        self.default_page = PageSession(
+            log=self,
+            log_formats=self.log_formats,
+            index_name=index_name,
+            target_dir=self.target_dir,
+            continuous_write=self.continuous_write,
+            log_to_stdout=self.log_to_stdout,
+            log_to_disk=log_to_disk,
+        )
+        """Defines the initial default target page in which the page data ia stored"""
+        self.default_builder: VisualLogBuilder = VisualLogBuilder(
+            self, page_session=self.default_page
+        )
         """
         The default builder. It let's you easily add content the log without
         the need of any callbacks.
@@ -447,7 +412,7 @@ class VisualLog:
         initialized with the flag `continuous=True` it can be terminated
         from within the logging function.
         """
-        with self._page_lock:
+        with self._general_lock:
             self._shall_terminate = True
 
     def add_static_file(self, filename: str, content: bytes):
@@ -460,7 +425,7 @@ class VisualLog:
         :param filename: The name of the data to add
         :param content: The file's content
         """
-        with self._page_lock:
+        with self._general_lock:
             self.static_files[filename] = content
 
     def get_file(self, filename: str) -> bytes | None:
@@ -471,7 +436,7 @@ class VisualLog:
         :param filename: The file's name
         :return: The file's content (if available)
         """
-        with self._page_lock:
+        with self._general_lock:
             if filename in self.static_files:
                 return self.static_files[filename]
             abs_filename = os.path.abspath(self.target_dir + "/" + filename)
@@ -502,12 +467,6 @@ class VisualLog:
         The maximum figure size in pixels
         """
         return Size2D(self._max_fig_size)
-
-    def handle_modified(self):
-        """
-        Is called when a new block of content has been inserted
-        """
-        pass
 
     def _provide_live_view(self):
         """
@@ -544,74 +503,6 @@ class VisualLog:
         )
         self.add_static_file("liveView.html", rendered_lv.encode("utf-8"))
 
-    def write_html(self, html_code: str | bytes):
-        """
-        The HTML code to add
-
-        :param html_code: The html code
-        :return: True if txt logging is enabled
-        """
-        if HTML not in self.log_formats:
-            return
-        if not isinstance(html_code, bytes):
-            html_code = html_code.encode("utf-8")
-        self.cur_element.add_data(HTML, html_code)
-        if self.continuous_write:
-            self.write_to_disk(formats={HTML})
-        return True
-
-    def write_md(self, md_code: str, no_break: bool = False):
-        """
-        The markdown code to add
-
-        :param md_code: The markdown code
-        :param no_break: If defined no line break will be added
-        :return: True if txt logging is enabled
-        """
-        if MD not in self.log_formats:
-            return
-        new_text = md_code + ("" if no_break else "\n")
-        self.cur_element.add_data(MD, new_text.encode("utf-8"))
-        if self.continuous_write:
-            self.write_to_disk(formats={MD})
-        return True
-
-    def write_txt(self, txt_code: str, console: bool = True, md: bool = False):
-        """
-        Adds text code to the txt / console log
-
-        :param txt_code: The text to add
-        :param console: Defines if the text shall also be added ot the
-            console's log (as it's mostly identical). True by default.
-        :param md: Defines if the text shall be added to markdown as well
-        :return: True if txt logging is enabled
-        """
-        if self.log_to_stdout:
-            print(txt_code)
-        if console and len(self._consoles):
-            self._add_to_console(txt_code)
-        if md and MD in self.log_formats:
-            self.write_md(txt_code)
-        if TXT not in self.log_formats:
-            return
-        self.cur_element.add_data(TXT, (txt_code + "\n").encode("utf-8"))
-        if self.continuous_write:
-            self.write_to_disk(formats={TXT})
-        return True
-
-    def _add_to_console(self, txt_code: str):
-        """
-        Adds text code to the console log
-
-        :param txt_code: The text to add
-        :return: True if txt logging is enabled
-        """
-        for console in self._consoles:
-            if console.progressive:
-                console.print(txt_code)
-        self.cur_element.add_data(CONSOLE, (txt_code + "\n").encode("ascii"))
-        return True
-
     def get_temp_path(self, relative: str | None = None) -> str:
         """
         Returns the temporary file path. The data will be wiped upon the call
@@ -626,232 +517,11 @@ class VisualLog:
             return FilePath.norm_path(self.tmp_path + "/" + relative)
         return self.tmp_path
 
-    def create_log_backup(self):
-        """
-        Creates a copy of the logs and stores it in the log backup
-        """
-        with self._page_lock:
-            log_copy = self._logs.clone()
-        with self._backup_lock:
-            self._log_backup = log_copy
-
-    def _build_body(self):
-        """
-        Requests to combine all logs and sub logs to a single page which
-        can be logged to the disk or provided in the browser. (excluding
-        html headers and footers), so just "the body" of the HTML page.
-
-        :return: The finalized page, e.g. by combining base_log w/
-            sub_logs as shown in the :class:`VisualLiveLog` class.
-        """
-        body: dict[str, bytes] = {}
-
-        for cur_format in self.log_formats:
-            log_data = self._logs.build(cur_format)
-
-            if cur_format == HTML:
-                body[cur_format] = self._renderers[HTML].build_body(log_data)
-            else:
-                body[cur_format] = log_data
-        return body
-
     def flush(self):
         """
         Writes the current state to disk
         """
-        self.write_to_disk()
-
-    def set_latest_page(self, page_type: str, content: bytes):
-        """
-        Stores a copy of the latest page.
-
-        This method is multi-threading secure.
-
-        :param page_type: The format of the page to store
-        :param content: The page's new content
-        """
-        with self._page_lock:
-            self._page_backups[page_type] = content
-
-    def get_page(self, format_type: str) -> bytes:
-        """
-        Receives the newest update of the page of given output type.
-
-        If not done automatically (e.g. when using a VisualLiveLog)
-        you might have to call render_pages yourself.
-
-        This method is multi-threading secure.
-
-        Assumes that render() or write_to_disk() or render was called before
-        since the last change. This is not necessary if continuous_write is
-        enabled.
-
-        :param format_type: The type of the page you want to receive
-        :return: The page's content.
-        """
-        with self._page_lock:
-            self.last_page_request = time.time()
-            if format_type in self._page_backups:
-                return self._page_backups[format_type]
-            return b""
-
-    def get_element(
-        self, name: str | None, output_format: str = HTML, backup: bool = False
-    ) -> (float, bytes):
-        """
-        Returns the element at given node, starting with the root element
-        ROOT_ELEMENT_NAME
-
-        :param name: The name of the element to return.
-
-            By default the root element will be returned.
-        :param output_format: The format to return
-        :param backup: Defines if the last backup shall be accessed rather than the
-            working data.
-        :return: Timestamp of the last update, the element's content
-        """
-        if name is None:
-            name = ROOT_ELEMENT_NAME
-        cleaned_name = name.replace("-", ".")
-        tree = cleaned_name.split(".")
-        if len(tree) == 0 or tree[0] != ROOT_ELEMENT_NAME:
-            return 0.0, b""
-
-        if backup:
-            access_lock = self._backup_lock
-            element: VisualLogElement = self._log_backup
-        else:
-            access_lock = self._page_lock
-            element: VisualLogElement = self._logs
-
-        with access_lock:
-            for cur_sub_name in tree[1:]:
-                if cur_sub_name in cur_sub_name:
-                    element = element[cur_sub_name]
-                else:
-                    return 0.0, b""
-
-            if element is None:
-                return 0.0, b""
-            latest_update = max(
-                element.last_child_update_time, element.last_direct_change_time
-            )
-            return latest_update, element.build(output_format)
-
-    def get_body(self, format_type: str) -> bytes:
-        """
-        Returns the latest body data.
-
-        Contains only the part of that format w/ header and footer.
-        Can be used to for example embed one log's content in another log
-        such as main_log.html(sub_log.get_body("html"))
-
-        Assumes that render() or write_to_disk() or render was called before
-        since the last change. This is not necessary if continuous_write is
-        enabled.
-
-        :param format_type: The type of the page you want to receive
-        :return: The page's content.
-        """
-        with self._page_lock:
-            if format_type in self._body_backups:
-                return self._body_backups[format_type]
-            return b""
-
-    def render(self, formats: set[str] | None = None) -> VisualLog:
-        """
-        Renders all pages - so combines the main log with the sub logs
-        of the supported output types (html, txt, md etc.) and stores
-        them.
-
-        The page data for each type can be received via :meth:`get_latest_page`.
-
-        :param formats: A set of the formats which shall be rendered.
-
-            None = All configured formats.
-        :return: The VisualLog object
-        """
-        if formats is None:
-            formats = self.log_formats
-        bodies = self._build_body()
-        with self._page_lock:
-            self._body_backups = bodies
-        # store html
-        if (
-            self._html_export
-            and self._html_filename is not None
-            and len(self._html_filename) > 0
-            and HTML in formats
-        ):
-            self.set_latest_page(HTML, self._renderers[HTML].build_page(bodies[HTML]))
-        # store markdown
-        if (
-            self.md_export
-            and self._md_filename is not None
-            and len(self._md_filename) > 0
-            and MD in formats
-        ):
-            self.set_latest_page(MD, bodies[MD])
-        # store txt
-        if (
-            self.txt_export
-            and self._txt_filename is not None
-            and len(self._txt_filename) > 0
-            and TXT in formats
-        ):
-            self.set_latest_page(TXT, bodies[TXT])
-        if CONSOLE in formats:
-            for console in self._consoles:
-                if console.progressive:
-                    continue
-                console.clear()
-                body = bodies[CONSOLE]
-                console.print(body.decode("ascii"))
-        return self
-
-    def write_to_disk(self, formats: set[str] | None = None, render=True) -> VisualLog:
-        """
-        Writes the rendered pages from all (or all specified) formats to
-        disk.
-
-        :param formats: A set of formats to write. None = all configured
-
-            e.g. {"html, "txt") etc. By default all formats will be stored.
-        :param render: Defines if the pages shall be rendered (if necessary)
-        :return: The VisualLog object
-        """
-        if formats is None:
-            formats = self.log_formats
-
-        if render:
-            self.render(formats=formats)
-
-        if self.log_to_disk:
-            # store html
-            if (
-                self._html_export
-                and self._html_filename is not None
-                and len(self._html_filename) > 0
-                and HTML in formats
-            ):
-                FileStag.save(self._html_filename, self.get_page(HTML))
-                # store markdown
-            if (
-                self.md_export
-                and self._md_filename is not None
-                and len(self._md_filename) > 0
-                and MD in formats
-            ):
-                FileStag.save(self._md_filename, self.get_page(MD))
-            # store txt
-            if (
-                self.txt_export
-                and self._txt_filename is not None
-                and len(self._txt_filename) > 0
-                and TXT in formats
-            ):
-                FileStag.save(self._txt_filename, self.get_page(TXT))
-        return self
+        self.default_page.write_to_disk()
 
     def finalize(self) -> VisualLog:
         """
@@ -859,7 +529,7 @@ class VisualLog:
 
         :return: The VisualLog object
         """
-        self.write_to_disk(render=True)
+        self.default_page.write_to_disk(render=True)
         if FilePath.exists(self.tmp_path):
             shutil.rmtree(self.tmp_path)
         return self
@@ -1012,7 +682,7 @@ class VisualLog:
         test = self.testing or test
         self.testing = test
         if builder is not None:
-            builder = self.prepare_builder(builder)
+            builder = self.prepare_builder(builder, self.default_page)
         self._auto_clear = auto_clear
         self._continuous_build = continuous
         self._builder_handler = builder
@@ -1104,7 +774,7 @@ class VisualLog:
                 if overwrite is True or not self.load_old_logs():
                     self._run_builder(builder)
                     self.handle_event_list()
-                    self.write_to_disk()
+                    self.default_page.write_to_disk()
         mt = mt and not test
         server.start(mt=mt, test=test)
         self._start_app_or_browser(real_log=self, url=self.local_live_url)
@@ -1126,7 +796,7 @@ class VisualLog:
                 if overwrite is True or not self.load_old_logs():
                     self._run_builder(self._builder_handler)
                     self.handle_event_list()
-                    self.write_to_disk()
+                    self.default_page.write_to_disk()
             if wait:
                 self.sleep()
 
@@ -1184,7 +854,7 @@ class VisualLog:
             could successfully be loaded, so that no run was required.
         """
         if builder is not None:
-            builder = self.prepare_builder(builder)
+            builder = self.prepare_builder(builder, self.default_page)
         self._auto_clear = auto_clear
         self._continuous_build = continuous
         self._builder_handler = builder
@@ -1219,14 +889,14 @@ class VisualLog:
                 if overwrite is True or not self.load_old_logs():
                     self._run_builder(builder)
                     self.handle_event_list()
-                    self.write_to_disk()
+                    self.default_page.write_to_disk()
                 else:
                     return False
         if continuous:
             auto_clear = auto_clear if auto_clear is not None else True
             self._run_continuous(auto_clear, builder)
         if self.log_to_disk and HTML in self.log_formats and self._start_browser:
-            self._start_app_or_browser(self, url=self._html_filename)
+            self._start_app_or_browser(self, url=self.default_page._html_filename)
             return True
 
     def _run_builder(self, builder: BuilderTypes | None = None):
@@ -1236,26 +906,26 @@ class VisualLog:
         :param builder: The builder to be called from rebuild the log
         """
         if builder is not None:
-            with self._page_lock:
+            with self._general_lock:
                 if getattr(builder, "build", None) is not None:
                     builder.build_page()
                 else:
                     builder(self.default_builder)
-            self.create_log_backup()
-        self.write_to_disk()
+        self.default_page.write_to_disk()
 
-    def prepare_builder(self, builder: BuilderTypes):
+    def prepare_builder(self, builder: BuilderTypes, page_session: PageSession):
         """
         Prepapres the builder to be used for this log
 
         :param builder: The build helper, either a function which fills the
             log or an ancestor of VisualLogBuilder implementing at least the
             build_body method to do the same.
+        :param page_session: Defines the target page to which the builder shall write
         :return: The prepared build object
         """
         if isinstance(builder, type):
             builder: Type[VisualLogBuilder] | VisualLogBuilder
-            builder = builder(log=self)
+            builder = builder(log=self, page_session=page_session)
             from .visual_log_builder import VisualLogBuilder
 
             if not isinstance(builder, VisualLogBuilder):
@@ -1273,7 +943,6 @@ class VisualLog:
         :param url: Defines the URL which shall be opened
         """
 
-        port = real_log.server.port if real_log.server is not None else 80
         if self._start_browser:
             import webbrowser
 
@@ -1349,14 +1018,14 @@ class VisualLog:
         while True:
             self._update_counter += 1
             self._total_update_counter += 1
-            with self._page_lock:
+            with self._general_lock:
                 if self._shall_terminate:
                     break
             if auto_clear:
                 self.clear()
             self._run_builder(builder)
             self.handle_event_list()
-            self.write_to_disk()
+            self.default_page.write_to_disk()
             cur_time = time.time()
             while cur_time < next_update:
                 time.sleep(next_update - cur_time)
@@ -1475,7 +1144,7 @@ class VisualLog:
 
         :param event: The new event
         """
-        with self._page_lock:
+        with self._general_lock:
             self._events.append(event)
             self.cache.set(LOG_EVENT_CACHE_NAME, self._events)
 
@@ -1483,7 +1152,7 @@ class VisualLog:
         """
         Handles all queued events and clears the event queue
         """
-        with self._page_lock:
+        with self._general_lock:
             event_list = self._events
             self._events = []
             self.cache.set(LOG_EVENT_CACHE_NAME, self._events)
@@ -1506,7 +1175,7 @@ class VisualLog:
         :param clear: Defines if all events shall be removed afterwards
         :return: The event list
         """
-        with self._page_lock:
+        with self._general_lock:
             event_list = list(self._events)
             if clear:
                 self._events = []
@@ -1571,24 +1240,11 @@ class VisualLog:
             uptime=time.time() - self.start_time,
         )
 
-    def embed(self, log_data: VisualLog):
-        """
-        Embeds another VisualLog's content into this one
-
-        :param log_data: The source log
-        """
-        for cur_format in self.log_formats:
-            if cur_format in log_data.log_formats:
-                self.cur_element.add_data(cur_format, log_data.get_body(cur_format))
-
     def clear(self):
         """
-        Clears the whole log (excluding headers and footers)
+        Clears the default page log (excluding headers and footers)
         """
-        self.name_counter = Counter()
-        self.title_counter = Counter()
-        self._logs.clear()
-        self.cur_element = self._logs
+        self.default_page.clear()
 
     @classmethod
     def is_main(cls) -> bool:
@@ -1639,20 +1295,6 @@ class VisualLog:
         :return: True if it is a mock
         """
         return False
-
-    def reserve_unique_name(self, name: str):
-        """
-        Reserves a unique name within the log, e.g. to store an object to
-        a unique file.
-
-        :param name: The desired name
-        :return: The effective name with which the data shall be stored
-        """
-        self.name_counter[name] += 1
-        result = name
-        if self.name_counter[name] > 1:
-            result += f"_{self.name_counter[name]}"
-        return result
 
     def __enter__(self) -> "VisualLogBuilder":
         """
